Refactored: Pulled out stand alone helper functions.
